feat(ups): full support and parameter decoding for APC UPS (US-040) (#10)
* feat: implement experimental APC UPS telemetry decoding (US-040)

* fix(apc): update APCDriver to poll and decode APC HID reports properly for BE600M1

* fix(apc): update configVoltage, extract missing data and fix inputVoltage bug

---------

+++ lib/USBHostUPS/src/USBHostUPS.h
@@ -33,14 +33,14 @@ struct UPSData {
 
     uint16_t configApparentPower = 0;
     uint8_t configFrequency = 0;
-    uint8_t configVoltage = 0;
+    uint16_t configVoltage = 0;
 
     uint16_t outputVoltage = 0;
     uint16_t inputVoltage = 0;
     uint16_t batteryVoltage = 0;
 
     uint16_t highVoltageTransfer = 0;
-    uint8_t lowVoltageTransfer = 0;
+    uint16_t lowVoltageTransfer = 0;
 
     String manufacturer = "";
     String product = "";
